conda_requirements.txt
# These are software packages necessary to run the WebGnomeAPI web server.
# This file is designed for use with conda:
#
# so these packages can be installed with the commands:
#
#     $ conda install --file conda_requirements.txt
#
# many of the requirements are also available via pip, but they may
# have slightly different names, but you can give it a try:
#
#     $ pip install -r conda_requirements.txt
#
# All required packages should be available on the conda-forge channel

# this is intended to be installed with py_gnome
#  -- python version should already be pinned
python>=3.9,<3.14

setuptools<81  # need old version for Pyramid and pkg_resources


# numpy is not strictly needed by the API, but is by gnome, etc,
# and adding it helps keep conda from trying to up- or down-grade it.

numpy
regex
awesome-slugify
pyramid>=2,<3
paste
pastedeploy>=2.0
pyramid_log
pyramid_session_redis
cornice>=3.4
ujson
docutils
htmlmin
jsmin
gitpython
waitress

pycparser
gevent
python-socketio
gevent-websocket

# With versions of redis-py greater than 4.6.0, we are getting the following:
#     ValueError: I/O operation on closed file.
#
# Reference: https://github.com/redis/redis-py/issues/3184
#
# Let's check this issue from time to time to see if we can un-pin the version
# on this package.
#
redis-py=4.6.0

# we need the redis server to manage sessions
# uncomment is you want to use a conda installed version
# redis

# This is to allow the web server to send e-mails from a google account,
# which we will be doing for the help feedback. 
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
